Restructuring tests to isolate the specialized template configurations for better management
