+++ tests/posit/specialized/CMakeLists.txt
@@ -1,0 +1,3 @@
+file (GLOB SOURCES "./*.cpp")
+
+compile_all("true" "fast" "${SOURCES}")
