+++ tests/posit/specialized/softposit16_ref.hpp
