+++ tests/posit/specialized/softposit32_ref.hpp
