+++ tests/posit/specialized/softposit8_ref.hpp
