+++ tests/posit/specialized/softposit_cmp.hpp
